+++ src/backend/chat-service/src/main/java/capstone/chatservice/domain/forum/controller/ForumMessageCommandController.java
@@ -7,7 +7,7 @@ import capstone.chatservice.domain.forum.dto.request.ForumMessageModifyRequest;
 import capstone.chatservice.domain.forum.service.command.ForumMessageCommandService;
 import capstone.chatservice.domain.model.UploadFile;
 import capstone.chatservice.infra.S3.FileStore;
-import capstone.chatservice.infra.kafka.KafkaProducer;
+import capstone.chatservice.infra.kafka.producer.KafkaProducer;
 import java.io.IOException;
 import java.util.List;
 import lombok.RequiredArgsConstructor;
