+++ src/backend/chat-service/src/main/java/capstone/chatservice/domain/server/controller/ServerMessageCommandController.java
@@ -7,7 +7,7 @@ import capstone.chatservice.domain.server.dto.request.ServerMessageDeleteRequest
 import capstone.chatservice.domain.server.dto.request.ServerMessageModifyRequest;
 import capstone.chatservice.domain.server.service.command.ServerMessageCommandService;
 import capstone.chatservice.infra.S3.FileStore;
-import capstone.chatservice.infra.kafka.KafkaProducer;
+import capstone.chatservice.infra.kafka.producer.KafkaProducer;
 import java.io.IOException;
 import java.util.List;
 import lombok.RequiredArgsConstructor;
