+++ src/backend/chat-service/src/main/java/capstone/chatservice/global/config/kafka/KafkaProducerConfig.java
@@ -4,8 +4,8 @@ import capstone.chatservice.domain.dm.dto.DirectMessageDto;
 import capstone.chatservice.domain.emoji.dto.EmojiDto;
 import capstone.chatservice.domain.forum.dto.ForumMessageDto;
 import capstone.chatservice.domain.server.dto.ServerMessageDto;
-import capstone.chatservice.infra.kafka.dto.ConnectionStateEventDto;
-import capstone.chatservice.infra.kafka.dto.ConnectionStateInfo;
+import capstone.chatservice.infra.kafka.producer.dto.ConnectionStateEventDto;
+import capstone.chatservice.infra.kafka.producer.dto.ConnectionStateInfo;
 import java.util.HashMap;
 import java.util.Map;
 import org.apache.kafka.common.serialization.StringSerializer;
