+++ src/backend/chat-service/src/main/java/capstone/chatservice/infra/client/StateServiceClient.java
@@ -1,7 +1,7 @@
 package capstone.chatservice.infra.client;
 
 import capstone.chatservice.global.common.dto.DataResponseDto;
-import capstone.chatservice.infra.kafka.dto.ConnectionStateInfo;
+import capstone.chatservice.infra.kafka.producer.dto.ConnectionStateInfo;
 import org.springframework.cloud.openfeign.FeignClient;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
