+++ src/backend/chat-service/src/main/java/capstone/chatservice/infra/kafka/producer/KafkaProducer.java
@@ -1,11 +1,11 @@
-package capstone.chatservice.infra.kafka;
+package capstone.chatservice.infra.kafka.producer;
 
 import capstone.chatservice.domain.dm.dto.DirectMessageDto;
 import capstone.chatservice.domain.emoji.dto.EmojiDto;
 import capstone.chatservice.domain.forum.dto.ForumMessageDto;
 import capstone.chatservice.domain.server.dto.ServerMessageDto;
-import capstone.chatservice.infra.kafka.dto.ConnectionStateEventDto;
-import capstone.chatservice.infra.kafka.dto.ConnectionStateInfo;
+import capstone.chatservice.infra.kafka.producer.dto.ConnectionStateEventDto;
+import capstone.chatservice.infra.kafka.producer.dto.ConnectionStateInfo;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Value;
