+++ src/backend/chat-service/src/main/java/capstone/chatservice/infra/kafka/producer/dto/ConnectionStateInfo.java
@@ -1,4 +1,4 @@
-package capstone.chatservice.infra.kafka.dto;
+package capstone.chatservice.infra.kafka.producer.dto;
 
 import lombok.Builder;
 import lombok.Getter;
