+++ src/backend/chat-service/src/main/java/capstone/chatservice/infra/websocket/WebSocketConnectionHandler.java
@@ -3,9 +3,9 @@ package capstone.chatservice.infra.websocket;
 import capstone.chatservice.infra.client.CommunityServiceClient;
 import capstone.chatservice.infra.client.StateServiceClient;
 import capstone.chatservice.infra.client.UserServerDmInfo;
-import capstone.chatservice.infra.kafka.KafkaProducer;
-import capstone.chatservice.infra.kafka.dto.ConnectionStateEventDto;
-import capstone.chatservice.infra.kafka.dto.ConnectionStateInfo;
+import capstone.chatservice.infra.kafka.producer.KafkaProducer;
+import capstone.chatservice.infra.kafka.producer.dto.ConnectionStateEventDto;
+import capstone.chatservice.infra.kafka.producer.dto.ConnectionStateInfo;
 import java.util.Objects;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
